+++ src/components/AppShell.tsx
@@ -1,8 +1,7 @@
-import { AppShell, Header, Aside } from '@mantine/core';
+import { AppShell, Aside } from '@mantine/core';
 import NewsComponent from './NewsComponent';
 
 import TodoList from './TodoList';
-import WeatherComponent from './WeatherComponent';
 
 function Shell() {
   return (
@@ -12,11 +11,6 @@ function Shell() {
         <Aside width={{ base: 550 }} height={800} p="xs">
           <NewsComponent />
         </Aside>
-      }
-      header={
-        <Header height={200} p="xs">
-          <WeatherComponent />
-        </Header>
       }
     >
       <TodoList />
